Move has_structural_eq_impls provider to rustc_ty_utils
